requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
numpy==1.24.3
openai==0.27.6
pvporcupine==2.2.0
PyAudio==0.2.13
pycparser==2.21
pydantic==1.10.7
python-dotenv==1.0.0
regex==2023.5.5
requests==2.30.0
tiktoken==0.3.3
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.2
yarl==1.9.2
